backend/video_pipeline/requirements.txt
# Core LLM
groq>=0.9.0

# PDF parsing
PyMuPDF>=1.23.0

# Text-to-Speech (free, no key needed)
edge-tts>=6.1.9

# Image generation (uses free Pollinations.ai API)
requests>=2.31.0
Pillow>=10.0.0

# Video assembly
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Web UI
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
tqdm>=4.66.0
pydantic>=2.0.0